+++ src/xpcs/filter/filter.h
@@ -74,9 +74,9 @@ public:
 
   virtual float* PartialPartitionsMean() = 0;
 
-  virtual float* TimestampClock() = 0;
+  virtual double* TimestampClock() = 0;
 
-  virtual float* TimestampTicks() = 0;
+  virtual double* TimestampTicks() = 0;
 
   virtual xpcs::data_structure::SparseData* Data() = 0;
